+++ hallo_eltern_cli/cli.py
@@ -42,6 +42,8 @@ def run():
     (args, config) = parse_arguments()
 
     _class = args.command if 'command' in args else DEFAULT_COMMAND_CLASS
+    if args.version:
+        _class = commands.VersionCommand
     command = _class(args, config)
     command.run()