+++ hallo_eltern_cli/__init__.py
@@ -4,7 +4,7 @@
 
 # flake8: noqa
 
-__version__ = '1.1.0-alpha.0'
+__version__ = '1.1.0'
 
 # Scaffolding functions
 from .common import get_config