+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Breaking Change: Split `mta` command into `stdout` and `mda` commands
+* Breaking Change: Require Python `>=3.7`
 * Enhancement: Add `smtp` command
 * Enhancement: Add `--force-address` to email generating commands
 * Enhancement: Add `--mda-command` to `mda` command